Switch to using REFERER
HURUmap dashboard is now a subdirectory-based multisite WP installation.
The use of HTTP_HOST will not work anymore since all sites will
essential be hosted on same domain (localhost:8000 when local and
dashboard.hurumap.org in PRODUCTION).

While not perfect, the use of REFERER header is the quickest and easiest
way to get working code while we figure out a fool proof plan